requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
pyyaml==6.0.1

# LangChain & MCP
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
langgraph==0.0.20
mcp==0.9.0

# Data Processing
pandas==2.1.4
numpy==1.26.2
yfinance==0.2.32
ta==0.11.0

# Async & HTTP
aiohttp==3.9.1
httpx==0.25.2
websockets==12.0
asyncio==3.4.3

# Database Clients
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1
chromadb==0.4.22

# TimescaleDB & SQL
sqlalchemy==2.0.23
sqlalchemy-utils==0.41.1

# API Clients
polygon-api-client==1.13.0
requests==2.31.0

# LLM & NLP
openai==1.6.1
tiktoken==0.5.2

# FastAPI & Web
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-multipart==0.0.6

# Utilities
python-dateutil==2.8.2
pytz==2023.3
beautifulsoup4==4.12.2
lxml==4.9.3

# Logging
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

